refactor the files to reflect the header file reorganization
